// ebk_front/home/bible/app/lib/config/interface.ts
export interface Eglise {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id_eglise: number;
  nom_eglise: string;
  username_eglise: string;
  photo_eglise: string;
  couverture_eglise: string;
  sigle_eglise: string;
  adresse_eglise: string;
  ville_eglise: string;
  pays_eglise: string;
  nombrefidel_eglise: string;
  status_eglise: string;
  payement_eglise: boolean;
  programme?: Programme[];
  word_of_day_eglise: string;
  about_eglise: string;
  localisation_eglise: string[];
}
// ...
export interface Programme {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  titre: string;
  sousProgramme: SousProgramme[];
}
// ...
export interface ManagementEvent {
  createdAt: Date;
  updatedAt: Date;
  deletedAt: null;
  id: number;
  name: string;
  description: string;
  dateEvent: Date;
  isBlocked: boolean;
  adressMap: string;
  totalPerson: number;
  isFree: boolean;
  price: number;
  isCancel: null;
  eglise: Eglise;
  totalSubscriptions: number;
  isSubscribe: boolean;
  annonces: string[];
  subscribe:
    | {
        createdAt: Date;
        id: number;
        uuid: string;
        isCancel: boolean;
        isChecked: boolean;
        paymentMothod: string;
        paymentReference: string;
      }
    | {
        createdAt: Date;
        id: number;
        uuid: string;
        isCancel: boolean;
        isChecked: boolean;
        paymentMothod: string;
        paymentReference: string;
      }[]
    | null;
}